15/8

+++ controllers/receiptController.ts
@@ -7,12 +7,19 @@ export class ReceiptController {
 	get = async (req: Request, res: Response) => {
 		try {
 			const userID = parseInt(req.params.id)
+
 			if (!userID) {
 				res.json({ success: false, message: "userID missing" })
 				return
 			}
+
+			if (isNaN(userID)) {
 				res.status(400).json({
 					message: 'Invalid ID'
+				})
+				return
+			}
+
 			const allReceipt = await this.receiptService.getReceipt(userID)
 			res.json(allReceipt)
 		} catch (err) {
@@ -22,6 +29,20 @@ export class ReceiptController {
 
 	post = async (req: Request, res: Response) => {
 		try {
+			if (
+				!req.params.id ||
+				!req.body.shopName ||
+				!req.body.date ||
+				!req.body.amount ||
+				!req.body.image ||
+				!req.body.expensesType
+			) {
+				res.status(400).json({
+					success: false,
+					message: 'Missing Information'
+				})
+				return
+			}
 			const userID = parseInt(req.params.id)
 			const receiptName = req.body.shopName
 			const receiptDate = req.body.date
@@ -30,13 +51,6 @@ export class ReceiptController {
 			const expensesType = req.body.expensesType
 			const is_deleted = false
 
-			if (!userID || !receiptName || !receiptDate || !receiptAmount || !receiptImage || !expensesType) {
-				res.json({
-					success: false,
-					message: 'Body Missing'
-				})
-				return
-			}
 
 			const result = await this.receiptService.addReceipt(
 				userID,
@@ -49,9 +63,9 @@ export class ReceiptController {
 			)
 
 			if (result) {
-				res.json({ success: true })
+				res.status(200).json({ success: true })
 			} else {
-				res.json({
+				res.status(500).json({
 					success: false,
 					message: 'Fail to save your receipt'
 				})
@@ -69,8 +83,14 @@ export class ReceiptController {
 			const revisedAmount = req.body.amount
 			const revisedType = req.body.type
 
-			if (!receiptID || !revisedVenue || !revisedDate || !revisedAmount || !revisedType) {
-				res.json({ success: false })
+			if (
+				!revisedVenue ||
+				!revisedDate ||
+				!revisedAmount ||
+				!revisedType ||
+				isNaN(receiptID)
+			) {
+				res.status(400).json({ success: false })
 				return
 			}
 
@@ -79,10 +99,11 @@ export class ReceiptController {
 				revisedVenue,
 				revisedDate,
 				revisedAmount,
-				revisedType)
+				revisedType
+			)
 
 			if (result.length === 0) {
-				res.json({ success: false })
+				res.status(400).json({ success: false })
 			} else {
 				res.json({ success: true })
 			}
@@ -95,28 +116,30 @@ export class ReceiptController {
 	delete = async (req: Request, res: Response) => {
 		try {
 			const receiptID = parseInt(req.params.id)
-			const result = await this.receiptService.deleteReceipt(receiptID)
-			if (result.length === 0) {
-				res.json({ success: false })
+			if (isNaN(receiptID)) {
+				res.status(400).json({ success: false })
+				return
 			}
-			res.json({ success: true })
-
+			const result = await this.receiptService.deleteReceipt(receiptID)
+			if (result.length === 0 || result.length > 1) {
+				res.status(500).json({ success: false })
+			}
+			res.status(200).json({ success: true })
 		} catch (err) {
 			console.log(err.message)
 		}
-
 	}
 
 	getSevenDay = async (req: Request, res: Response) => {
 		try {
 			const userID = req.params.id
-			if (!userID) {
-				res.json({ message: 'No params ID' })
+			if (!userID || isNaN(parseInt(userID))) {
+				res.status(400).json({ message: 'No params ID' })
 				return
 			}
 			const sevenDaysReceipt =
 				await this.receiptService.getSevenDaysReceipt(userID)
-			res.json(sevenDaysReceipt)
+			res.status(200).json(sevenDaysReceipt)
 		} catch (err) {
 			console.log(err.message)
 		}
@@ -125,13 +148,13 @@ export class ReceiptController {
 	getMonthly = async (req: Request, res: Response) => {
 		try {
 			const userID = req.params.id
-			if (!userID) {
-				res.json({ message: 'No params ID' })
+			if (!userID || isNaN(parseInt(userID))) {
+				res.status(400).json({ message: 'No params ID' })
 				return
 			}
 			const monthlyTypeReceipt =
 				await this.receiptService.getReceiptByThisMonth(userID)
-			res.json(monthlyTypeReceipt)
+			res.status(200).json(monthlyTypeReceipt)
 		} catch (e) {
 			console.log(e.message)
 		}
